+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 11.7
 - Remove useless center option in ListOutput() & program wide
 - Fix SQL when "Search All Schools" checked in GetTeacher.fnc.php
 - Add `_makeHonorRollGPAMinInputs()` function, group Honor Roll GPA Min inputs in same column in ReportCardGrades.php
+- Add "No Grades found for Percent" error in InputFinalGrades.php
 
 Changes in 11.6
 ---------------